deleted:    Jenkinsfile 	deleted:    Steps.md 	deleted:    _img/image-20220827003903489.png 	modified:   ansible/playbooks/roles/deploy/tasks/main.yml 	renamed:    ansible/playbooks/roles/deploy/template/goldenwaste.service -> ansible/playbooks/roles/deploy/template/goldenwaste.service.j2 	modified:   ansible/playbooks/roles/docker/tasks/main.yml 	modified:   ansible/playbooks/roles/git/tasks/main.yml 	modified:   ansible/playbooks/roles/mysql/tasks/main.yml 	modified:   ansible/playbooks/roles/nginx/tasks/files/site-enable.sh 	modified:   ansible/playbooks/roles/nginx/tasks/main.yml 	modified:   ansible/playbooks/roles/setup1/tasks/main.yml 	renamed:    ansible/playbooks/roles/setup1/templates/config -> ansible/playbooks/roles/setup1/templates/config.j2 	renamed:    ansible/playbooks/roles/setup1/templates/credentials -> ansible/playbooks/roles/setup1/templates/credentials.j2 	modified:   ansible/playbooks/roles/setup2/tasks/main.yml 	renamed:    ansible/playbooks/roles/setup2/templates/application.properties -> ansible/playbooks/roles/setup2/templates/application.properties.j2 	deleted:    hackathon.json

+++ ansible/playbooks/roles/nginx/tasks/files/site-enable.sh
@@ -1,22 +1,25 @@
-public_ip=$(curl ipinfo.io/ip)
-echo $public_ip
+if test -f "/etc/nginx/sites-available/goldenwaste"; then
+    echo "SKIPPED"
+else
+    public_ip=$(curl ipinfo.io/ip)
 
-echo '
-server {
-    listen 80;
+    echo '
+    server {
+        listen 80;
 
-    server_name public_ip;
+        server_name public_ip;
 
-    location / {
-        proxy_pass http://localhost:8080;
-        proxy_http_version 1.1;
-        proxy_set_header Upgrade $http_upgrade;
-        proxy_set_header Connection 'upgrade';
-        proxy_set_header Host $host;
-        proxy_cache_bypass $http_upgrade;
-    }
-}' > goldenwaste
+        location / {
+            proxy_pass http://localhost:8080;
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection 'upgrade';
+            proxy_set_header Host $host;
+            proxy_cache_bypass $http_upgrade;
+        }
+    }' > goldenwaste
 
-sed -i 's/public_ip/'${public_ip}'/g' goldenwaste
-sudo cp goldenwaste /etc/nginx/sites-available/goldenwaste
-ln -s /etc/nginx/sites-available/goldenwaste /etc/nginx/sites-enabled
+    sed -i 's/public_ip/'${public_ip}'/g' goldenwaste
+    sudo cp goldenwaste /etc/nginx/sites-available/goldenwaste
+    ln -s /etc/nginx/sites-available/goldenwaste /etc/nginx/sites-enabled
+fi
